// frontend/src/api/generated/models/catalogAppInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Palmimo Portal
 * OpenAPI spec version: 0.1.0
 */
import type { AppSourceInfo } from './appSourceInfo';
import type { EnvSpecInfo } from './envSpecInfo';

export interface CatalogAppInfo {
  description: string;
  devices: string[];
  env: EnvSpecInfo[];
  name: string;
  source: AppSourceInfo;
}